requirements.txt
ell-ai[all]
google-genai
mistralai
datasets
polars
jax
joblib